Use unordered_map for TreeContainer

+++ src/openrct2/paint/PaintStructDescriptor.cpp
@@ -202,16 +202,16 @@ void PaintStructDescriptor::Paint(
     }
 }
 
-std::vector<size_t> PaintStructKeyInserter::GetParams(const PaintStructDescriptorKey& key) const
+std::vector<uint32_t> PaintStructKeyInserter::GetParams(const PaintStructDescriptorKey& key) const
 {
-    return std::vector<size_t>{ key.Direction,           key.Element,
+    return std::vector<uint32_t>{ key.Direction,           key.Element,
                                 key.TrackSequence,       key.VehicleKey[0].NumPeeps,
                                 key.VehicleKey[0].Pitch, key.VehicleKey[0].SpriteDirection };
 }
 
 // don't put the track sequence for the image id
-std::vector<size_t> ImageIdKeyInserter::GetParams(const PaintStructDescriptorKey& key) const
+std::vector<uint32_t> ImageIdKeyInserter::GetParams(const PaintStructDescriptorKey& key) const
 {
-    return std::vector<size_t>{ key.Direction, key.Element, key.VehicleKey[0].NumPeeps, key.VehicleKey[0].Pitch,
+    return std::vector<uint32_t>{ key.Direction, key.Element, key.VehicleKey[0].NumPeeps, key.VehicleKey[0].Pitch,
                                 key.VehicleKey[0].SpriteDirection };
 }
